+++ changelog.md
@@ -3,3 +3,4 @@
 - Made currently open file show up in the window title.
 - Status bar will show cursor grid coordinates when nothing else is showing.
 - Tile cursor will stay in place while using the scroll wheel if the mouse is not moved.
+- Textures and shapes copied from other tiles will be highlighted properly in the texture/shape picker modes.